[#48] Button extends BaseComponent
